+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Coinbase is no longer a supported payment method. `PAYMENT_METHOD_NO_LONGER_SUPPORTED` will be returned for Coinbase operations.
 * Add facilitated transaction details to Transaction if present
 * Add `bin` to `ApplePayCard`
+* Add `submitForSettlement` to `SubscriptionGateway.retryCharge`
 
 ## 2.72.1